+++ src/view/home.js
@@ -1,27 +1,12 @@
 /* eslint-disable import/no-unresolved */
-import {
-  createPost, auth, getPost, getUser, deletePost, editPost,
-} from '../firebase/auth.js';
+import viewPost from './post.js';
+import viewProfile from './profile.js';
 
 export default () => {
   // CREACIÓN DEL TEMPLATE
   const viewHome = `
     <section class="container-profile">
       <div class="profile-card">
-        <div class="img-card-profile">
-          <img src="img/perfilwhite.png" alt="imagen de perfil">
-        </div>
-        <div class="user-name">
-          <h5>${auth.currentUser.email}</h5>
-        </div>
-        <div class="user-info">
-          <div class="about-user">
-            <span class="about-me">Mis juegos favoritos son...</span>
-          </div>
-          <div class="more-info">
-            <span class"more-about-me">Interacción</span>
-          </div>
-        </div>
       </div>
       <div class="info-dev">
         <span>Desarrollado por</span>
@@ -66,124 +51,10 @@ export default () => {
   const section = document.createElement('section');
   section.setAttribute('class', 'home-page');
   section.innerHTML = viewHome;
+  const divProfile = section.querySelector('.profile-card');
 
-  // DECLARACION DE CONSTANTES PARA MANEJO DEL DOM
-  const btnSharePost = section.querySelector('#btn-share-post');
-  const containerPostPublicated = section.querySelector('.container-post');
-
-  // EVENTO CLICK DEL BOTON COMPARTIR EL POST
-  btnSharePost.addEventListener('click', () => {
-    // METODO PARA OBTENER LA FECHA Y HORA EN LA QUE SE REALIZA EL POSTEO
-    const date = new Date();
-    const datePost = `${(date.getDate()).toString().padStart(2, '0')}/${(date.getMonth() + 1).toString().padStart(2, '0')}/${date.getFullYear()} ${date.getHours().toString().padStart(2, '0')}:${date.getMinutes().toString().padStart(2, '0')}:${date.getSeconds().toString().padStart(2, '0')}`;
-    createPost(
-      auth.currentUser.uid,
-      section.querySelector('#create-post').textContent,
-      datePost,
-      'público',
-    );
-    section.querySelector('#create-post').textContent = '';
-  });
-
-  // FUNCIÓN PARA MOSTRAR LOS POST
-  const querySnapshot = (queryPost) => {
-    containerPostPublicated.innerHTML = '';
-    queryPost.forEach((post) => {
-      // OBTENEMOS LA INFORMACIÓN DEL USUARIO DUEÑO DEL POST
-      getUser(post.data().uid).then((user) => {
-        const userData = user.data();
-        const userName = userData.name;
-        const divPostPublicated = document.createElement('div');
-        divPostPublicated.setAttribute('class', 'container-publicated');
-        divPostPublicated.innerHTML = `
-        <div class='post-publicated'>
-        <div class='info-user'>
-          <div class='info-post'>
-            <div class='photo-perfil-post'>
-              <img src='img/perfilblack.png' alt='foto perfil de usuario'>
-            </div>
-            <div class='nameuser-date'>
-              <span>${userName}</span> <br>
-              <span>${post.data().datePost}</span>
-            </div>
-          </div>
-          <div class='btn-edit-delete'>
-            <i class='bi bi-three-dots'></i>
-          </div>
-        </div>
-        <div class='input-readonly'>
-          <span id=${post.id} class='post-publicated cont-post' role='textbox'>${post.data().post}</span>
-        </div>
-      </div>
-      <div class='container-like-comment'>
-        <div class='mando-img'>
-          <i class='bi bi-joystick'></i>
-          <span class='span-counter'></span>
-          <span class='span-text'> Me gusta </span>
-        </div>
-        <div class='comment-img'>
-          <i class='bi bi-chat-dots'></i>
-          <span class='span-text'> Comentar </span>
-        </div>
-        <div>
-          <button type='button' class='btn-edit'>Editar</button> 
-        </div>
-        <div>
-          <button type='button' class='btn-delete'>Eliminar</button> 
-        </div>
-        <div>
-          <button type='button' class='btn-save'>Guardar</button> 
-        </div>
-        </div>`;
-
-        containerPostPublicated.appendChild(divPostPublicated);
-
-        // TRAEMOS LOS BOTONES DE ELIMINAR Y EDITAR LOS POST
-        const btnDelete = divPostPublicated.querySelector('.btn-delete');
-        const btnEdit = divPostPublicated.querySelector('.btn-edit');
-        const btnSave = divPostPublicated.querySelector('.btn-save');
-        const contentPost = document.getElementById(`${post.id}`);
-        const btnLike = divPostPublicated.querySelector('.bi-joystick');
-        let count = 0;
-        btnLike.addEventListener('click', () => {
-          console.log('like');
-          count += 1;
-          btnLike.classList.toggle('active-like');
-          divPostPublicated.querySelector('.span-counter').textContent = count;
-        });
-        // EVENTO CLICK PARA ELIMINAR LOS POST
-        btnDelete.addEventListener('click', () => {
-          const opcion = window.confirm('¿Estás seguro que deseas eliminar el post?');
-          if (opcion === true) {
-            deletePost(post.id).then(() => {
-              console.log('post eliminado ', post.id);
-            });
-          } else {
-            console.log('Eliminación cancelada');
-          }
-        });
-        // EVENTOS DE CLICK PARA EDITAR LOS POST
-        btnEdit.addEventListener('click', () => {
-          contentPost.setAttribute('contentEditable', 'true');
-          contentPost.focus();
-        });
-        btnSave.addEventListener('click', () => {
-          const opcion = window.confirm('¿Estás seguro que deseas guardar los cambios?');
-          if (opcion === true) {
-            editPost(post.id, {
-              post: contentPost.textContent,
-            }).then(() => {
-              contentPost.removeAttribute('contentEditable');
-              console.log('post editado');
-            });
-          } else {
-            console.log('Edición cancelada');
-          }
-        });
-      });
-    });
-  };
-  getPost(querySnapshot);
+  viewProfile(divProfile);
+  viewPost(section);
 
   return section; // RETORNA EL NODO DE LA SECCION DE HOME
 };
